refactor: move async job component in its own package
