+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/async_job/query_service/AsyncJobQueryService.java
@@ -13,17 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.apim.core.integration.crud_service;
+package io.gravitee.apim.core.async_job.query_service;
 
-import io.gravitee.apim.core.integration.model.AsyncJob;
+import io.gravitee.apim.core.async_job.model.AsyncJob;
 import java.util.Optional;
 
-public interface AsyncJobCrudService {
-    AsyncJob create(AsyncJob job);
-
-    Optional<AsyncJob> findById(String id);
-
-    AsyncJob update(AsyncJob job);
-
-    void delete(String id);
+public interface AsyncJobQueryService {
+    Optional<AsyncJob> findPendingJobFor(String integrationId);
 }
